+++ lib/status-page/services/redis.rb
@@ -5,10 +5,24 @@ module StatusPage
     class RedisException < StandardError; end
 
     class Redis < Base
+      class Configuration
+        attr_accessor :url
+
+        def initialize
+          @url = "redis://127.0.0.1:3306/1"
+        end
+      end
+
+      class << self
+        def config_class
+          Redis::Configuration
+        end
+      end
+
       def check!
         time = Time.now.to_s(:db)
 
-        redis = ::Redis.current
+        redis = ::Redis.new(url: config.url)
         redis.set(key, time)
         fetched = redis.get(key)
 
@@ -16,7 +30,7 @@ module StatusPage
       rescue Exception => e
         raise RedisException.new(e.message)
       ensure
-        redis.client.disconnect
+        redis.disconnect
       end
 
       private