+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.1.4
+
+- Allow Redis to config host url.
+- Update service options API in configure.
+
 0.1.3
 -----